+++ packages/twofinger/src/Q4C12/TwoFinger/Internal.hs
@@ -44,8 +44,6 @@ import Data.Stream.Infinite
 import qualified Data.Stream.Infinite as Stream
 import Data.Traversable (foldMapDefault, fmapDefault)
 import GHC.Generics (Generic)
-import Test.QuickCheck (Gen)
-import qualified Test.QuickCheck as QC
 
 -- $setup
 -- >>> import Data.List (unfoldr)
@@ -61,8 +59,6 @@ import qualified Test.QuickCheck as QC
 --TODO: Alternative zippy Applicatives instances.
 
 --TODO: Consider exporting bits and pieces from, e.g., Q4C12.TwoFinger.EvenA, without the flavour-identifying suffix, to allow qualified import.
-
---TODO: The prop> lines are very long due to a doctest limiation: https://github.com/sol/doctest/issues/131. When this is fixed, I should make those reasonable.
 
 --TODO: the issue with the mathy haddocks is that double-clicking on a paragraph with one of them in them won't select the whole paragraph.
 
@@ -513,8 +509,6 @@ unsnocOddA tree = case first halfunsnocEvenA $ halfunsnocOddA tree of
   (Just (tree', e), a) -> Right (tree', (e, a))
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfunconsEvenE'
---
--- prop> \e (AnyOddA as) -> halfunconsEvenE (halfconsOddA e as) == Just (e, as)
 halfconsOddA :: e -> TwoFingerOddA e a -> TwoFingerEvenE e a
 halfconsOddA e (EmptyOddA a) = SingleEvenE e a
 halfconsOddA e (SingleOddA a1 e1 a2) =
@@ -524,8 +518,6 @@ halfconsOddA e (DeepOddA a0 pr m sf a1) = case digitCons e a0 pr of
   Left (pr', a', node) -> DeepEvenE pr' (consOddA a' node m) sf a1
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfunsnocEvenA'
---
--- prop> \(AnyOddA as) e -> halfunsnocEvenA (halfsnocOddA as e) == Just (as, e)
 halfsnocOddA :: TwoFingerOddA e a -> e -> TwoFingerEvenA e a
 halfsnocOddA (EmptyOddA a) e = SingleEvenA a e
 halfsnocOddA (SingleOddA a e1 a1) e2 =
@@ -535,16 +527,12 @@ halfsnocOddA (DeepOddA a0 pr m sf a1) e = case digitSnoc sf a1 e of
   Left (node, a', sf') -> DeepEvenA a0 pr (snocOddA m node a') sf'
 
 -- | \(O(1)\) worst case. Inverse: 'halfconsEvenE'
---
--- prop> \(AnyOddA as) -> as == uncurry halfconsEvenE (halfunconsOddA as)
 halfunconsOddA :: TwoFingerOddA e a -> (a, TwoFingerEvenE e a)
 halfunconsOddA (EmptyOddA a) = (a, EmptyEvenE)
 halfunconsOddA (SingleOddA a e1 a1) = (a, SingleEvenE e1 a1)
 halfunconsOddA (DeepOddA a0 pr m sf a1) = (a0, DeepEvenE pr m sf a1)
 
 -- | \(O(1)\) worst case. Inverse: 'halfsnocOddA'
---
--- prop> \(AnyOddA as) -> as == uncurry halfsnocEvenA (halfunsnocOddA as)
 halfunsnocOddA :: TwoFingerOddA e a -> (TwoFingerEvenA e a, a)
 halfunsnocOddA (EmptyOddA a) = (EmptyEvenA, a)
 halfunsnocOddA (SingleOddA a1 e1 a2) = (SingleEvenA a1 e1, a2)
@@ -568,22 +556,16 @@ unsnocOddE tree = case first halfunsnocEvenE $ halfunsnocOddE tree of
   (Just (tree', a), e) -> Right (tree', (a, e))
 
 -- | \(O(1)\) worst case. Inverse: 'halfunconsEvenA'
---
--- prop> \a (AnyOddE as) -> halfunconsEvenA (halfconsOddE a as) == Just (a, as)
 halfconsOddE :: a -> TwoFingerOddE e a -> TwoFingerEvenA e a
 halfconsOddE a (SingleOddE e) = SingleEvenA a e
 halfconsOddE a (DeepOddE pr m sf) = DeepEvenA a pr m sf
 
 -- | \(O(1)\) worst case. Inverse: 'halfunsnocEvenE'
---
--- prop> \(AnyOddE as) a -> halfunsnocEvenE (halfsnocOddE as a) == Just (as, a)
 halfsnocOddE :: TwoFingerOddE e a -> a -> TwoFingerEvenE e a
 halfsnocOddE (SingleOddE e) a = SingleEvenE e a
 halfsnocOddE (DeepOddE pr m sf) a = DeepEvenE pr m sf a
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfconsEvenA'
---
--- prop> \(AnyOddE as) -> as == uncurry halfconsEvenA (halfunconsOddE as)
 halfunconsOddE :: TwoFingerOddE e a -> (e, TwoFingerEvenA e a)
 halfunconsOddE (SingleOddE e) = (e, EmptyEvenA)
 halfunconsOddE (DeepOddE pr m sf) = case digitUncons pr of
@@ -591,8 +573,6 @@ halfunconsOddE (DeepOddE pr m sf) = case digitUncons pr of
   (e, Just (a, pr')) -> (e, DeepEvenA a pr' m sf)
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfsnocEvenE'
---
--- prop> \(AnyOddE as) -> as == uncurry halfsnocEvenE (halfunsnocOddE as)
 halfunsnocOddE :: TwoFingerOddE e a -> (TwoFingerEvenE e a, e)
 halfunsnocOddE (SingleOddE e) = (EmptyEvenE, e)
 halfunsnocOddE (DeepOddE pr m sf) = case digitUnsnoc sf of
@@ -617,16 +597,12 @@ unsnocEvenE tree = case first halfunsnocOddE <$> halfunsnocEvenE tree of
   Just ((tree', a), e) -> Just (tree', (a, e))
 
 -- | \(O(1)\) worst case. Inverse: 'halfunconsOddA'
---
--- prop> \a (AnyEvenE as) -> halfunconsOddA (halfconsEvenE a as) == (a, as)
 halfconsEvenE :: a -> TwoFingerEvenE e a -> TwoFingerOddA e a
 halfconsEvenE a EmptyEvenE = EmptyOddA a
 halfconsEvenE a0 (SingleEvenE e1 a1) = SingleOddA a0 e1 a1
 halfconsEvenE a0 (DeepEvenE pr m sf a1) = DeepOddA a0 pr m sf a1
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfunsnocOddE'.
---
--- prop> \(AnyEvenE as) e -> halfunsnocOddE (halfsnocEvenE as e) == (as, e)
 halfsnocEvenE :: TwoFingerEvenE e a -> e -> TwoFingerOddE e a
 halfsnocEvenE EmptyEvenE e = SingleOddE e
 halfsnocEvenE (SingleEvenE e1 a1) e2 =
@@ -636,8 +612,6 @@ halfsnocEvenE (DeepEvenE pr m sf a) e = case digitSnoc sf a e of
   Left (node, a', sf') -> DeepOddE pr (snocOddA m node a') sf'
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfconsOddA'.
---
--- prop> \(AnyEvenE as) -> as == maybe mempty (uncurry halfconsOddA) (halfunconsEvenE as)
 halfunconsEvenE :: TwoFingerEvenE e a -> Maybe (e, TwoFingerOddA e a)
 halfunconsEvenE EmptyEvenE = Nothing
 halfunconsEvenE (SingleEvenE e a) = Just (e, EmptyOddA a)
@@ -646,8 +620,6 @@ halfunconsEvenE (DeepEvenE pr m sf a1) = Just $ case digitUncons pr of
   (e, Just (a0, pr')) -> (e, DeepOddA a0 pr' m sf a1)
 
 -- | \(O(1)\) worst case. Inverse: 'halfsnocOddE'.
---
--- prop> \(AnyEvenE as) -> as == maybe mempty (uncurry halfsnocOddE) (halfunsnocEvenE as)
 halfunsnocEvenE :: TwoFingerEvenE e a -> Maybe (TwoFingerOddE e a, a)
 halfunsnocEvenE EmptyEvenE = Nothing
 halfunsnocEvenE (SingleEvenE e a) = Just (SingleOddE e, a)
@@ -671,8 +643,6 @@ unsnocEvenA tree = case first halfunsnocOddA <$> halfunsnocEvenA tree of
   Just ((tree', e), a) -> Just (tree', (e, a))
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfunconsOddE'.
---
--- prop> \e (AnyEvenA as) -> halfunconsOddE (halfconsEvenA e as) == (e, as)
 halfconsEvenA :: e -> TwoFingerEvenA e a -> TwoFingerOddE e a
 halfconsEvenA e EmptyEvenA = SingleOddE e
 halfconsEvenA e1 (SingleEvenA a1 e2) =
@@ -682,24 +652,18 @@ halfconsEvenA e (DeepEvenA a pr m sf) = case digitCons e a pr of
   Left (pr', a', node) -> DeepOddE pr' (consOddA a' node m) sf
 
 -- | \(O(1)\) worst case. Inverse: 'halfunsnocOddA'.
---
--- prop> \(AnyEvenA as) a -> halfunsnocOddA (halfsnocEvenA as a) == (as, a)
 halfsnocEvenA :: TwoFingerEvenA e a -> a -> TwoFingerOddA e a
 halfsnocEvenA EmptyEvenA a = EmptyOddA a
 halfsnocEvenA (SingleEvenA a1 e1) a2 = SingleOddA a1 e1 a2
 halfsnocEvenA (DeepEvenA a0 pr m sf) a = DeepOddA a0 pr m sf a
 
 -- | \(O(1)\) worst case. Inverse: 'halfconsOddE'.
---
--- prop> \(AnyEvenA as) -> as == maybe mempty (uncurry halfconsOddE) (halfunconsEvenA as)
 halfunconsEvenA :: TwoFingerEvenA e a -> Maybe (a, TwoFingerOddE e a)
 halfunconsEvenA EmptyEvenA = Nothing
 halfunconsEvenA (SingleEvenA a e) = Just (a, SingleOddE e)
 halfunconsEvenA (DeepEvenA a pr m sf) = Just (a, DeepOddE pr m sf)
 
 -- | \(O(\log n)\) worst case. Inverse: 'halfsnocOddA'.
---
--- prop> \(AnyEvenA as) -> as == maybe mempty (uncurry halfsnocOddA) (halfunsnocEvenA as)
 halfunsnocEvenA :: TwoFingerEvenA e a -> Maybe (TwoFingerOddA e a, e)
 halfunsnocEvenA EmptyEvenA = Nothing
 halfunsnocEvenA (SingleEvenA a e) = Just (EmptyOddA a, e)
@@ -709,7 +673,6 @@ halfunsnocEvenA (DeepEvenA a1 pr m sf) = case digitUnsnoc sf of
 
 -- * Monad and Applicative instances, and related operations
 
---TODO: should be able to write some property tests for this.
 joinOddA :: TwoFingerOddA (TwoFingerOddE e a) (TwoFingerOddA e a) -> TwoFingerOddA e a
 joinOddA (halfunconsOddA -> (a, tree)) = appendOddAEvenE a (joinEvenE tree)
 
@@ -778,14 +741,9 @@ singletonOddE = SingleOddE
 
 -- * Concatenation of TwoFingerOddA.
 
--- |
--- prop> \(AnyOddA a) (AnyOddA b) (AnyOddA c) -> (a <> b) <> c == a <> (b <> c)
 instance (Semigroup a) => Semigroup (TwoFingerOddA e a) where
   (<>) = appendOddA0
 
--- |
--- prop> \(AnyOddA a) -> a == mempty <> a
--- prop> \(AnyOddA a) -> a == a <> mempty
 instance (Monoid a, Semigroup a) => Monoid (TwoFingerOddA e a) where
   mempty = singletonOddA mempty
   mappend = (<>)
@@ -1124,17 +1082,12 @@ addDigits4 m1 (Four e1 a1 e2 a2 e3 a3 e4) a4 e5 a5 e6 a6 e7 a7 e8 a8
 
 -- * Concatenation of TwoFingerEvenE.
 
--- |
--- prop> \(AnyEvenE a) (AnyEvenE b) (AnyEvenE c) -> (a <> b) <> c == a <> (b <> c)
 instance Semigroup (TwoFingerEvenE e a) where
   (<>) = appendEvenE
 
 instance Alt (TwoFingerEvenE e) where
   (<!>) = appendEvenE
 
--- |
--- prop> \(AnyEvenE a) -> a == a <> mempty
--- prop> \(AnyEvenE a) -> a == mempty <> a
 instance Monoid (TwoFingerEvenE e a) where
   mempty = EmptyEvenE
   mappend = (<>)
@@ -1152,17 +1105,12 @@ appendEvenE (DeepEvenE pr1 m1 sf1 b1) (DeepEvenE pr2 m2 sf2 b2) =
 
 -- * Concatenation of TwoFingerEvenA.
 
--- |
--- prop> \(AnyEvenA a) (AnyEvenA b) (AnyEvenA c) -> (a <> b) <> c == a <> (b <> c)
 instance Semigroup (TwoFingerEvenA e a) where
   (<>) = appendEvenA
 
 instance Alt (TwoFingerEvenA e) where
   (<!>) = appendEvenA
 
--- |
--- prop> \(AnyEvenA a) -> a == a <> mempty
--- prop> \(AnyEvenA a) -> a == mempty <> a
 instance Monoid (TwoFingerEvenA e a) where
   mempty = EmptyEvenA
   mappend = (<>)
@@ -1180,8 +1128,6 @@ appendEvenA (DeepEvenA a1 pr1 m1 sf1) (DeepEvenA a2 pr2 m2 sf2) =
 
 -- * Monoid actions
 
--- |
--- prop> \(AnyOddA a) -> a == appendOddAEvenE a mempty
 appendOddAEvenE :: TwoFingerOddA e a -> TwoFingerEvenE e a -> TwoFingerOddA e a
 appendOddAEvenE (EmptyOddA a) m = halfconsEvenE a m
 appendOddAEvenE m EmptyEvenE = m
@@ -1190,8 +1136,6 @@ appendOddAEvenE m (SingleEvenE e a) = snocOddA m e a
 appendOddAEvenE (DeepOddA a1 pr1 m1 sf1 a2) (DeepEvenE pr2 m2 sf2 a3) =
   DeepOddA a1 pr1 (addDigits0 m1 sf1 a2 pr2 m2) sf2 a3
 
--- |
--- prop> \(AnyOddA a) -> a == appendEvenAOddA mempty a
 appendEvenAOddA :: TwoFingerEvenA e a -> TwoFingerOddA e a -> TwoFingerOddA e a
 appendEvenAOddA EmptyEvenA m = m
 appendEvenAOddA m (EmptyOddA a) = halfsnocEvenA m a
@@ -1214,8 +1158,6 @@ appendOddEOddA m (SingleOddA a1 e a2) = snocEvenE (halfsnocOddE m a1) e a2
 appendOddEOddA (DeepOddE pr1 m1 sf1) (DeepOddA a1 pr2 m2 sf2 a2) =
   DeepEvenE pr1 (addDigits0 m1 sf1 a1 pr2 m2) sf2 a2
 
--- |
--- prop> \(AnyOddE a) -> a == appendOddEEvenA a mempty
 appendOddEEvenA :: TwoFingerOddE e a -> TwoFingerEvenA e a -> TwoFingerOddE e a
 appendOddEEvenA m EmptyEvenA = m
 appendOddEEvenA (SingleOddE e) m = halfconsEvenA e m
@@ -1223,106 +1165,12 @@ appendOddEEvenA m (SingleEvenA a e) = snocOddE m a e
 appendOddEEvenA (DeepOddE pr1 m1 sf1) (DeepEvenA a pr2 m2 sf2) =
   DeepOddE pr1 (addDigits0 m1 sf1 a pr2 m2) sf2
 
--- |
--- prop> \(AnyOddE a) -> a == appendEvenEOddE mempty a
 appendEvenEOddE :: TwoFingerEvenE e a -> TwoFingerOddE e a -> TwoFingerOddE e a
 appendEvenEOddE EmptyEvenE m = m
 appendEvenEOddE (SingleEvenE a e) m = consOddE a e m
 appendEvenEOddE m (SingleOddE e) = halfsnocEvenE m e
 appendEvenEOddE (DeepEvenE pr1 m1 sf1 a) (DeepOddE pr2 m2 sf2) =
   DeepOddE pr1 (addDigits0 m1 sf1 a pr2 m2) sf2
-
--- * QuickCheck stuff.
-genDigit :: Gen e -> Gen a -> Gen (Digit e a)
-genDigit e a = QC.oneof
-  [ One <$> e
-  , Two <$> e <*> a <*> e
-  , Three <$> e <*> a <*> e <*> a <*> e
-  , Four <$> e <*> a <*> e <*> a <*> e <*> a <*> e
-  ]
-
-genNode :: Gen e -> Gen a -> Gen (Node e a)
-genNode e a = QC.oneof
-  [ Node2 <$> e <*> a <*> e
-  , Node3 <$> e <*> a <*> e <*> a <*> e
-  ]
-
--- | The 'Int' parameter is expontential size: for a value \(n\), the generated tree will have (slightly more than) \(2^n\) to \(3^n\) elements.
-genOddA :: Gen e -> Gen a -> Int -> Gen (TwoFingerOddA e a)
-genOddA e a 1 = SingleOddA <$> a <*> e <*> a
-genOddA _ a n | n <= 0 = EmptyOddA <$> a
-genOddA e a n =
-  DeepOddA <$> a <*> genDigit e a <*> genOddA (genNode e a) a (n - 2) <*> genDigit e a <*> a
-
---TODO: better shrinks? This isn't wrong, and it's better than the default, but we could be doing better (e.g., trying just the middle tree in Deep; also possibly just dropping things off the ends...).
-shrinkOddA :: TwoFingerOddA e a -> [TwoFingerOddA e a]
-shrinkOddA = \case
-  EmptyOddA _ -> []
-  SingleOddA a1 _ a2 ->
-    [ EmptyOddA a1
-    , EmptyOddA a2
-    ]
-  DeepOddA a0 pr m sf a1 -> mconcat
-    [ [ halfsnocEvenA (halfconsOddE a0 $ digitToTree pr) (fst $ halfunconsOddA m)
-      , halfconsEvenE (snd $ halfunsnocOddA m) (halfsnocOddE (digitToTree sf) a1)
-      ]
-    , [EmptyOddA a0]
-    , [EmptyOddA a1]
-    , (\m' -> DeepOddA a0 pr m' sf a1) <$> shrinkOddA m
-    ]
-
-shrinkOddE :: TwoFingerOddE e a -> [TwoFingerOddE e a]
-shrinkOddE (SingleOddE _) = []
-shrinkOddE (DeepOddE pr m sf) = (\m' -> DeepOddE pr m' sf) <$> shrinkOddA m
-
-shrinkEvenA :: TwoFingerEvenA e a -> [TwoFingerEvenA e a]
-shrinkEvenA tree = case unconsEvenA tree of
-  Nothing -> []
-  Just (_, tree') -> [tree']
-
-shrinkEvenE :: TwoFingerEvenE e a -> [TwoFingerEvenE e a]
-shrinkEvenE tree = case unconsEvenE tree of
-  Nothing -> []
-  Just (_, tree') -> [tree']
-
-newtype AnyOddA = AnyOddA { getAnyOddA :: TwoFingerOddA Int [Int] }
-  deriving (Show)
-
-instance QC.Arbitrary AnyOddA where
-  arbitrary = fmap AnyOddA $ genOddA QC.arbitrary QC.arbitrary =<< QC.choose (0, 10)
-  shrink = fmap AnyOddA . shrinkOddA . getAnyOddA
-
-newtype AnyOddE = AnyOddE { getAnyOddE :: TwoFingerOddE Int [Int] }
-  deriving (Show)
-
-instance QC.Arbitrary AnyOddE where
-  arbitrary = AnyOddE <$> QC.oneof
-    [ SingleOddE <$> QC.arbitrary
-    , DeepOddE <$> genDigit QC.arbitrary QC.arbitrary <*> (genOddA (genNode QC.arbitrary QC.arbitrary) QC.arbitrary =<< QC.choose (0, 10)) <*> genDigit QC.arbitrary QC.arbitrary
-    ]
-  shrink = fmap AnyOddE . shrinkOddE . getAnyOddE
-
-newtype AnyEvenA = AnyEvenA { getAnyEvenA :: TwoFingerEvenA Int [Int] }
-  deriving (Show)
-
-instance QC.Arbitrary AnyEvenA where
-  arbitrary = AnyEvenA <$> QC.oneof
-    [ pure EmptyEvenA
-    , SingleEvenA <$> QC.arbitrary <*> QC.arbitrary
-    , DeepEvenA <$> QC.arbitrary <*> genDigit QC.arbitrary QC.arbitrary <*> (genOddA (genNode QC.arbitrary QC.arbitrary) QC.arbitrary =<< QC.choose (0, 10)) <*> genDigit QC.arbitrary QC.arbitrary
-    ]
-  shrink = fmap AnyEvenA . shrinkEvenA . getAnyEvenA
-
-newtype AnyEvenE = AnyEvenE { getAnyEvenE :: TwoFingerEvenE Int [Int] }
-  deriving (Show)
-
-instance QC.Arbitrary AnyEvenE where
-  arbitrary = AnyEvenE <$> QC.oneof
-    [ pure EmptyEvenE
-    , SingleEvenE <$> QC.arbitrary <*> QC.arbitrary
-    , DeepEvenE <$> genDigit QC.arbitrary QC.arbitrary <*> (genOddA (genNode QC.arbitrary QC.arbitrary) QC.arbitrary =<< QC.choose (0, 10)) <*> genDigit QC.arbitrary QC.arbitrary <*> QC.arbitrary
-    ]
-  shrink = fmap AnyEvenE . shrinkEvenE . getAnyEvenE
 
 -- * Aligning/zipping.
 
@@ -1333,8 +1181,6 @@ instance QC.Arbitrary AnyEvenE where
 --
 -- >>> alignLeftOddAOddA (consOddA 'a' 1 $ singletonOddA 'b') (consOddA "foo" 10 $ consOddA "bar" 20 $ singletonOddA "baz")
 -- (consOddA ('a',"foo") (1,10) (singletonOddA ('b',"bar")),Right (consEvenE 20 "baz" mempty))
---
--- prop> \(AnyOddA as) (AnyOddA bs) -> let { (aligned, rest) = alignLeftOddAOddA as bs ; as' = appendOddAEvenE (bimap fst fst aligned) (either id (const mempty) rest) ; bs' = appendOddAEvenE (bimap snd snd aligned) (either (const mempty) id rest) } in as == as' && bs == bs'
 alignLeftOddAOddA :: TwoFingerOddA e a -> TwoFingerOddA e' a' -> (TwoFingerOddA (e, e') (a, a'), Either (TwoFingerEvenE e a) (TwoFingerEvenE e' a'))
 alignLeftOddAOddA as (halfunsnocOddA -> (bs, a')) = case alignLeftOddAEvenA as bs of
   Left (aligned, halfunconsOddA -> (a, rest)) ->
@@ -1348,8 +1194,6 @@ alignLeftOddAOddA as (halfunsnocOddA -> (bs, a')) = case alignLeftOddAEvenA as b
 --
 -- >>> alignLeftOddAEvenA (consOddA 'a' 1 $ singletonOddA 'b') (consEvenA "foo" 10 $ consEvenA "bar" 20 $ consEvenA "baz" 30 mempty)
 -- Right (consOddA ('a',"foo") (1,10) (singletonOddA ('b',"bar")),consOddE 20 "baz" (singletonOddE 30))
---
--- prop> \(AnyOddA as) (AnyEvenA bs) -> let { (as', bs') = case alignLeftOddAEvenA as bs of { Left (aligned, rest) -> (appendEvenAOddA (bimap fst fst aligned) rest, bimap snd snd aligned) ; Right (aligned, rest) -> (bimap fst fst aligned, appendOddAOddE (bimap snd snd aligned) rest) } } in as == as' && bs == bs'
 alignLeftOddAEvenA :: TwoFingerOddA e a -> TwoFingerEvenA e' a' -> Either (TwoFingerEvenA (e, e') (a, a'), TwoFingerOddA e a) (TwoFingerOddA (e, e') (a, a'), TwoFingerOddE e' a')
 alignLeftOddAEvenA as bs = case (unconsOddA as, unconsEvenA bs) of
   (Right ((a, e), as'), Just ((a', e'), bs')) -> case alignLeftOddAEvenA as' bs' of
@@ -1371,15 +1215,11 @@ alignLeftEvenAEvenA as bs = case (unconsEvenA as, unconsEvenA bs) of
 --
 -- >>> alignLeftOddEOddE (consOddE 'a' 1 $ singletonOddE 'b') (consOddE "foo" 10 $ consOddE "bar" 20 $ singletonOddE "baz")
 -- (consOddE ('a',"foo") (1,10) (singletonOddE ('b',"bar")),Right (consEvenA 20 "baz" mempty))
---
--- prop> \(AnyOddE as) (AnyOddE bs) -> let { (aligned, rest) = alignLeftOddEOddE as bs ; as' = appendOddEEvenA (bimap fst fst aligned) (either id (const mempty) rest) ; bs' = appendOddEEvenA (bimap snd snd aligned) (either (const mempty) id rest) } in as == as' && bs == bs'
 alignLeftOddEOddE :: TwoFingerOddE e a -> TwoFingerOddE e' a' -> (TwoFingerOddE (e, e') (a, a'), Either (TwoFingerEvenA e a) (TwoFingerEvenA e' a'))
 alignLeftOddEOddE as (halfunsnocOddE -> (bs, e')) = case alignLeftOddEEvenE as bs of
   Left (aligned, halfunconsOddE -> (e, rest)) -> (halfsnocEvenE aligned (e, e'), Left rest)
   Right (aligned, rest) -> (aligned, Right $ halfsnocOddA rest e')
 
--- |
--- prop> \(AnyOddE as) (AnyEvenE bs) -> let { (as', bs') = case alignLeftOddEEvenE as bs of { Left (aligned, rest) -> (appendEvenEOddE (bimap fst fst aligned) rest, bimap snd snd aligned) ; Right (aligned, rest) -> (bimap fst fst aligned, appendOddEOddA (bimap snd snd aligned) rest) } } in as == as' && bs == bs'
 alignLeftOddEEvenE :: TwoFingerOddE e a -> TwoFingerEvenE e' a' -> Either (TwoFingerEvenE (e, e') (a, a'), TwoFingerOddE e a) (TwoFingerOddE (e, e') (a, a'), TwoFingerOddA e' a')
 alignLeftOddEEvenE as bs = case (unconsOddE as, unconsEvenE bs) of
   (Right ((e, a), as'), Just ((e', a'), bs')) -> case alignLeftOddEEvenE as' bs' of
@@ -1430,9 +1270,6 @@ infiniteOddA' f g (a0 :> leftA) leftE rightE (an :> rightA) =
   in DeepOddA a0 (nodeToDigit prNode) inner (nodeToDigit sfNode) an
 
 -- | Infinitely repeat the given @a@ and @e@.
---
--- prop> \(AnyOddA as) -> as == bimap (uncurry ($)) (uncurry ($)) (fst $ alignLeftOddAOddA (repeatOddA id id) as)
--- prop> \(AnyEvenA as) -> either ((as ==) . bimap (uncurry ($)) (uncurry ($)) . fst) (const False) (alignLeftOddAEvenA (repeatOddA id id) as)
 repeatOddA :: a -> e -> TwoFingerOddA e a
 repeatOddA a e = infiniteOddA (Stream.iterate id a) (Stream.iterate id e) (Stream.iterate id e) (Stream.iterate id a)
 
@@ -1448,9 +1285,6 @@ infiniteOddA :: Stream a -> Stream e -> Stream e -> Stream a -> TwoFingerOddA e 
 infiniteOddA = infiniteOddA' takeSingleNodeLeft takeSingleNodeRight
 
 -- | Infinitely repeat the given @a@ and @e@.
---
--- prop> \(AnyOddE as) -> as == bimap (uncurry ($)) (uncurry ($)) (fst $ alignLeftOddEOddE (repeatOddE id id) as)
--- prop> \(AnyEvenE as) -> either ((==) as . bimap (uncurry ($)) (uncurry ($)) . fst) (const False) $ alignLeftOddEEvenE (repeatOddE id id) as
 repeatOddE :: e -> a -> TwoFingerOddE e a
 repeatOddE e a = infiniteOddE (Stream.iterate id e) (Stream.iterate id a) (Stream.iterate id a) (Stream.iterate id e)
 
@@ -1470,9 +1304,6 @@ infiniteOddE leftE leftA rightA rightE =
     inner = infiniteOddA' (takeNodeLeft takeSingleNodeLeft) (takeNodeRight takeSingleNodeRight) leftE' leftA' rightA' rightE'
 
 -- | Infinitely repeat the given @a@ and @e@.
---
--- prop> \(AnyEvenA as) -> as == bimap (uncurry ($)) (uncurry ($)) (fst $ alignLeftEvenAEvenA (repeatEvenA id id) as)
--- prop> \(AnyOddA as) -> either (const False) ((==) as . bimap (uncurry $ flip ($)) (uncurry $ flip ($)) . fst) $ alignLeftOddAEvenA as (repeatEvenA id id)
 repeatEvenA :: a -> e -> TwoFingerEvenA e a
 repeatEvenA a e = infiniteEvenA (Stream.iterate id a) (Stream.iterate id e) (Stream.iterate id a) (Stream.iterate id e)
 
@@ -1487,10 +1318,6 @@ infiniteEvenA :: Stream a -> Stream e -> Stream a -> Stream e -> TwoFingerEvenA 
 infiniteEvenA (a :> leftA) leftE rightA rightE =
   halfconsOddE a $ infiniteOddE leftE leftA rightA rightE
 
--- |
---
--- prop> \(AnyEvenE as) -> as == bimap (uncurry ($)) (uncurry ($)) (fst $ alignLeftEvenEEvenE (repeatEvenE id id) as)
--- prop> \(AnyOddE as) -> either (const False) ((==) as . bimap (uncurry $ flip ($)) (uncurry $ flip ($)) . fst) $ alignLeftOddEEvenE as (repeatEvenE id id)
 repeatEvenE :: e -> a -> TwoFingerEvenE e a
 repeatEvenE e a = infiniteEvenE (Stream.iterate id e) (Stream.iterate id a) (Stream.iterate id e) (Stream.iterate id a)